Feat: Chatgpt (#2)
* chatgpt build

* feat: add .gitignore to exclude ai-commit and .env files

This commit introduces a .gitignore file to prevent tracking of
the 'ai-commit' and '.env' files in the repository, ensuring
sensitive information and temporary files remain untracked.

* chore: update workflow template version to 0.0.7

This update changes the container.yaml version used in the build workflow.
No other changes have been made to the workflow configuration.

* feat: add proxy support to HTTP client in callOpenAI function

Enhanced the HTTP client configuration to include proxy settings from the environment. This change allows for better flexibility in network configurations when making API calls.